# lib/api-zod/src/generated/types/addressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoleHer Women's Shoe Store API
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInput {
  label: string;
  street: string;
  city: string;
  country: string;
  postalCode: string;
  isDefault?: boolean;
}