# lib/api-zod/src/generated/types/catalogueStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoleHer Women's Shoe Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogueStats {
  brands: string[];
  styles: string[];
  heelHeights: string[];
  availableSizes: number[];
  colours: string[];
  minPrice: number;
  maxPrice: number;
}